requirements.txt
# backend/requirements.txt

Flask
Flask-CORS
PyJWT
bcrypt
Werkzeug
python-dotenv
gunicorn         
psycopg2-binary 
decimal          